finished new slippi logic
